requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0
djangorestframework==3.13.1
idna==3.3
pika==1.2.0
psycopg2==2.9.3
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
